+++ config/configtest/configtest.go
@@ -32,7 +32,7 @@ import (
 // The regular expression for valid config field tag.
 var configFieldTagRegExp = regexp.MustCompile("^[a-z0-9][a-z0-9_]*$")
 
-// LoadConfig loads a config from file, and does NOT validate the configuration.
+// LoadConfig loads a config.Config  from file, and does NOT validate the configuration.
 func LoadConfig(fileName string, factories component.Factories) (*config.Config, error) {
 	// Read yaml config from file
 	cp, err := configmapprovider.NewExpand(configmapprovider.NewFile(fileName)).Retrieve(context.Background(), nil)
@@ -54,6 +54,16 @@ func LoadConfigAndValidate(fileName string, factories component.Factories) (*con
 		return nil, err
 	}
 	return cfg, cfg.Validate()
+}
+
+// LoadConfigMap loads a config.Map from file, and does NOT validate the configuration.
+func LoadConfigMap(fileName string) (*config.Map, error) {
+	ret, err := configmapprovider.NewFile(fileName).Retrieve(context.Background(), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return ret.Get(context.Background())
 }
 
 // CheckConfigStruct enforces that given configuration object is following the patterns